Override user_params whitelisting for Panama
Alaveteli has added a method to the users controllers to whitelist the
params that can be altered when signing up a new user. This stops our
custom user fields from being saved and so breaks the signup process.

This commit overrides the whitelisting method so that we can pass
these fields again.

Closes #90